set the players object id

+++ src/world.cpp
@@ -4,8 +4,8 @@
 #include "object/object.h"
 #include "object/object_attack.h"
 #include "object/character.h"
-#include "util/load_exception.h"
 #include "factory/object_factory.h"
+#include "util/load_exception.h"
 #include "globals.h"
 #include "object/effect.h"
 #include "object/enemy.h"
@@ -52,6 +52,9 @@ takeAScreenshot(false){
 
         for ( vector<PlayerTracker>::iterator it = this->players.begin(); it != this->players.end(); it++ ){
             PlayerTracker & tracker = *it;
+            /* players need a new id because the level could create objects with
+             * ids that overlap with the player
+             */
             tracker.player->setObjectId(ObjectFactory::getNextObjectId());
             void * handle = Script::Engine::getEngine()->createPlayer(tracker.player);
             tracker.player->setScriptObject(handle);
@@ -85,6 +88,7 @@ World::~World(){
             delete *it;
         }
 	
+        /* this prevents the object from being destroyed in the Object destructor */
         for ( vector<PlayerTracker>::iterator it = players.begin(); it != players.end(); it++ ){
             PlayerTracker & tracker = *it;
             void * handle = tracker.script;
@@ -244,7 +248,7 @@ void World::addEnemy(Enemy * obj){
 Object * World::findObject(int id){
     for (vector<Object*>::iterator it = objects.begin(); it != objects.end(); it++){
         Object * object = *it;
-        if (object->getObjectId() == (unsigned int) id){
+        if (object->getObjectId() == id){
             return object;
         }
     }
